more fixes
