+++ front/components/navbar.php
@@ -1,9 +1,9 @@
 <?php
-    if ($_SESSION['logged']) {
+if ($_SESSION['logged']) {
 
-        $userId = $_SESSION['user_id'];
+    $userId = $_SESSION['user_id'];
 
-        $_SESSION['btn'] = "
+    $_SESSION['btn'] = "
         <a href='/front/pages/profile?id=$userId' style='color:white;'> <i class='fas fa-user fa-lg fa-fw' style='line-height: 1.5;'></i></a>
 
         <a><button class='nav-item btn btn-link' type='submit' name='logout'
@@ -12,15 +12,15 @@
         <path fill-rule='evenodd' d='M6 12.5a.5.5 0 0 0 .5.5h8a.5.5 0 0 0 .5-.5v-9a.5.5 0 0 0-.5-.5h-8a.5.5 0 0 0-.5.5v2a.5.5 0 0 1-1 0v-2A1.5 1.5 0 0 1 6.5 2h8A1.5 1.5 0 0 1 16 3.5v9a1.5 1.5 0 0 1-1.5 1.5h-8A1.5 1.5 0 0 1 5 12.5v-2a.5.5 0 0 1 1 0v2z'/>
         <path fill-rule='evenodd' d='M.146 8.354a.5.5 0 0 1 0-.708l3-3a.5.5 0 1 1 .708.708L1.707 7.5H10.5a.5.5 0 0 1 0 1H1.707l2.147 2.146a.5.5 0 0 1-.708.708l-3-3z'/>
         </svg> </button></a>";
-        
-    }else {
-        $_SESSION['btn'] = "<a href='/front/pages/login' class='nav-item'>
+
+} else {
+    $_SESSION['btn'] = "<a href='/front/pages/login' class='nav-item'>
         <svg xmlns='http://www.w3.org/2000/svg' width='20' height='20' fill='currentColor' class='bi bi-box-arrow-in-right' viewBox='0 0 16 16'>
             <path fill-rule='evenodd' d='M6 3.5a.5.5 0 0 1 .5-.5h8a.5.5 0 0 1 .5.5v9a.5.5 0 0 1-.5.5h-8a.5.5 0 0 1-.5-.5v-2a.5.5 0 0 0-1 0v2A1.5 1.5 0 0 0 6.5 14h8a1.5 1.5 0 0 0 1.5-1.5v-9A1.5 1.5 0 0 0 14.5 2h-8A1.5 1.5 0 0 0 5 3.5v2a.5.5 0 0 0 1 0v-2z'/>
             <path fill-rule='evenodd' d='M11.854 8.354a.5.5 0 0 0 0-.708l-3-3a.5.5 0 1 0-.708.708L10.293 7.5H1.5a.5.5 0 0 0 0 1h8.793l-2.147 2.146a.5.5 0 0 0 .708.708l3-3z'/>
         </svg> Login</a>";
 
-    }
+}
 ?>
 <nav class="navbar navbar-expand-lg bg-body-tertiary sticky-top">
     <div class="container-fluid">
@@ -33,6 +33,12 @@
                 <li class="nav-item">
                     <a class="nav-link" href="/">Home</a>
                 </li>
+                <li class="nav-item">
+                    <span class="mx-2">•</span>
+                </li>
+                <li class="nav-item">
+                    <a class="nav-link" href="../editTypes.php">Listagem</a>
+                </li>
             </ul>
         </div>
 
@@ -44,9 +50,8 @@
         <div class="collapse navbar-collapse justify-content-center" id="navbarContent">
             <!-- Desktop Version -->
             <div class="d-none d-lg-flex align-items-center justify-content-between flex-grow-1">
-                <form class="d-flex mx-auto" role="search" style="width: 50%;" 
-                    action="/api/operations/audioVisualOperation" method='post'
-                >
+                <form class="d-flex mx-auto" role="search" style="width: 50%;"
+                    action="/api/operations/audioVisualOperation" method='post'>
                     <input class="form-control me-2" type="search" placeholder="Faça sua busca aqui..."
                         aria-label="Search" style="width: 100%;" name="search">
                     <button class="btn btn-outline-success" name="searchLike" type="submit">Buscar</button>
@@ -56,10 +61,10 @@
             <div class="d-none d-lg-flex align-items-center">
                 <form action="/api/operations/userOperation.php" method="post">
                     <?php
-                        if (isset($_SESSION['btn'])) {
-                            echo $_SESSION['btn'];
-                        }
-                    ?>  
+                    if (isset($_SESSION['btn'])) {
+                        echo $_SESSION['btn'];
+                    }
+                    ?>
                 </form>
             </div>
 
@@ -93,5 +98,5 @@
 </nav>
 
 <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.1/dist/js/bootstrap.bundle.min.js"
-        integrity="sha384-HwwvtgBNo3bZJJLYd8oVXjrBZt8cqVSpeBNS5n7C8IVInixGAoxmnlMuBnhbgrkm"
-        crossorigin="anonymous"></script>
+    integrity="sha384-HwwvtgBNo3bZJJLYd8oVXjrBZt8cqVSpeBNS5n7C8IVInixGAoxmnlMuBnhbgrkm"
+    crossorigin="anonymous"></script>
